feat(compount_templates): remove compound templates, they are assumptions of library use and can be created by users downstream when required
