+++ v0.1/src/sheetcake.ts
@@ -1,45 +1,10 @@
 // brian taylor vann
 // sheetcake
 
-import {
-  keyframe,
-  createAttributeSelector,
-  createMediaQuery,
-  createSelector,
-  style,
-} from "./template_functions/template_functions.ts";
-
-import {
-  checked,
-  disabled,
-  focus,
-  hover,
-  invalid,
-  landscape,
-  portrait,
-  print,
-  required,
-  screen1008,
-  screen641,
-  valid,
-} from "./compound_templates/compound_templates.ts";
-
 export {
   keyframe,
   createAttributeSelector,
   createMediaQuery,
   createSelector,
   style,
-  checked,
-  disabled,
-  focus,
-  hover,
-  invalid,
-  landscape,
-  portrait,
-  print,
-  required,
-  screen1008,
-  screen641,
-  valid,
-};
+} from "./template_functions/template_functions.ts";
